requirements.txt
# Core dependencies
opencv-python==4.8.0
face-recognition==1.3.0
mysql-connector-python==8.0.33
numpy==1.24.3

# Required by face-recognition
dlib==19.24.0
Pillow==10.0.0

# Development tools
pytest==7.4.0
black==23.7.0
flake8==6.1.0

# Optional but recommended
python-dotenv==1.0.0  # For environment variables
tqdm==4.65.0  # For progress bars